## global o/c rework ##
- removed unused/unnecessary computations/code from the global observer
- added a new GlobalController based around a modular configuration of it's functions
  - each task that the global controller performs was split into a modules that can be expanded and modified based on future needs
  - modules will be configured via the configuration files
  - modules will have a simple pub/sub system for events emitted by them
  - modules will share a data-storage for exchange of information
- added configuration parameters for an enforced scheduling delay at the start of the simulation and between each scheduling
- adjusted the o/c configuration file for better access to contents

+++ osh_comdriver_logging/src_sim/osh/comdriver/logging/EnergyLoggingComDriver.java
@@ -172,14 +172,14 @@ public class EnergyLoggingComDriver extends CALComDriver implements IDataRegistr
 
                 for (int i = 0; i < avgWeekDays.length; i++) {
                     for (int j = 0; j < avgWeekDays[i].length; j++) {
-                        double factor = (double) this.h0ResultsCounter[i][j] * PhysicalConstants.factor_wsToKWh;
+                        double factor = this.h0ResultsCounter[i][j] * PhysicalConstants.factor_wsToKWh;
                         for (int k = 0; k < avgWeekDays[i][j].length; k++) {
                             avgWeekDays[i][j][k] = this.aggrH0ResultsWeekdays[i][j][k] / factor;
                         }
                     }
                 }
                 for (int i = 0; i < avgDays.length; i++) {
-                    double factor = ((double) this.h0ResultsCounterDays[i] / 1440.0) * PhysicalConstants.factor_wsToKWh;
+                    double factor = (this.h0ResultsCounterDays[i] / 1440.0) * PhysicalConstants.factor_wsToKWh;
                     for (int j = 0; j < avgDays[i].length; j++) {
                         avgDays[i][j] = this.aggrH0ResultsDays[i][j] / factor;
                     }
